chore: unifying apis, and moving schemas to dedicated file
